Time: 1234 ms (33.33%), Space: 191.5 MB (16.67%) - LeetHub

+++ 2477-minimum-fuel-cost-to-report-to-the-capital/2477-minimum-fuel-cost-to-report-to-the-capital.cpp
@@ -35,8 +35,7 @@ public:
         }
         
         ans = 0LL;
-        
-        long long curr  = DFS(0,seats,-1,adj);
+        int curr  = DFS(0,seats,-1,adj);
     
         
         return ans;
